Send to Email changes...
Modds to client and server.  The Quote Summary page always has to act like it's in 'from the recent list' mode otherwise the redirect bacl from Cancel on any of the Merge pages breaks.  The asp page sends back a Key2 (person) value to ennable the email addresss to be correctly determined